fixup! Remove PipelineRunner#apply
